refactor: Refactor picker generate to multiple files (#22850)
* move single picker out

* move generateRangePicker out

* fix compile
